Moved unit tests into respective unit test files
