+++ test_cpp_generator.py
@@ -1,7 +1,5 @@
 import unittest
-import filecmp
 import os
-import io
 
 from code_generator import *
 from cpp_generator import *
@@ -10,159 +8,6 @@ __doc__ = """
 Unit tests for C++ code generator
 """
 
-
-def handle_to_factorial(_, cpp):
-    cpp('return n < 1 ? 1 : (n * factorial(n - 1));')
-
-
-class TestCppFunctionGenerator(unittest.TestCase):
-
-    @staticmethod
-    def handle_to_factorial(_, cpp):
-        cpp('return n < 1 ? 1 : (n * factorial(n - 1));')
-
-    def test_is_constexpr_raises_error_when_implementation_value_is_none(self):
-        writer = io.StringIO()
-        cpp = CppFile(None, writer=writer)
-        func = CppFunction(name="factorial", ret_type="int", is_constexpr=True)
-        self.assertRaises(RuntimeError, func.render_to_string, cpp)
-
-    def test_is_constexpr_render_to_string(self):
-        writer = io.StringIO()
-        cpp = CppFile(None, writer=writer)
-        func = CppFunction(name="factorial", ret_type="int",
-                           implementation_handle=TestCppFunctionGenerator.handle_to_factorial, is_constexpr=True)
-        func.add_argument('int n')
-        func.render_to_string(cpp)
-        self.assertIn(dedent("""\
-            constexpr int factorial(int n)
-            {
-            \treturn n < 1 ? 1 : (n * factorial(n - 1));
-            }"""), writer.getvalue())
-
-    def test_is_constexpr_render_to_string_declaration(self):
-        writer = io.StringIO()
-        cpp = CppFile(None, writer=writer)
-        func = CppFunction(name="factorial", ret_type="int",
-                           implementation_handle=TestCppFunctionGenerator.handle_to_factorial, is_constexpr=True)
-        func.add_argument('int n')
-        func.render_to_string_declaration(cpp)
-        self.assertIn(dedent("""\
-            constexpr int factorial(int n)
-            {
-            \treturn n < 1 ? 1 : (n * factorial(n - 1));
-            }"""), writer.getvalue())
-
-    def test_README_example(self):
-        writer = io.StringIO()
-        cpp = CppFile(None, writer=writer)
-        factorial_function = CppFunction(name='factorial', ret_type='int', is_constexpr=True,
-                                         implementation_handle=handle_to_factorial,
-                                         documentation='/// Calculates and returns the factorial of p n.')
-        factorial_function.add_argument('int n')
-        factorial_function.render_to_string(cpp)
-        self.assertIn(dedent("""\
-            /// Calculates and returns the factorial of p n.
-            constexpr int factorial(int n)
-            {
-            \treturn n < 1 ? 1 : (n * factorial(n - 1));
-            }"""), writer.getvalue())
-
-
-class TestCppVariableGenerator(unittest.TestCase):
-
-    def test_cpp_var_via_writer(self):
-        writer = io.StringIO()
-        cpp = CppFile(None, writer=writer)
-        variables = CppVariable(name="var1",
-                                type="char*",
-                                is_class_member=False,
-                                is_static=False,
-                                is_const=True,
-                                initialization_value='0')
-        variables.render_to_string(cpp)
-        self.assertEqual('const char* var1 = 0;\n', writer.getvalue())
-
-    def test_is_constexpr_raises_error_when_is_const_true(self):
-        self.assertRaises(RuntimeError, CppVariable, name="COUNT", type="int", is_class_member=True, is_const=True,
-                          is_constexpr=True, initialization_value='0')
-
-    def test_is_constexpr_raises_error_when_initialization_value_is_none(self):
-        self.assertRaises(RuntimeError, CppVariable, name="COUNT", type="int", is_class_member=True, is_constexpr=True)
-
-    def test_is_constexpr_render_to_string(self):
-        writer = io.StringIO()
-        cpp = CppFile(None, writer=writer)
-        variables = CppVariable(name="COUNT",
-                                type="int",
-                                is_class_member=False,
-                                is_constexpr=True,
-                                initialization_value='0')
-        variables.render_to_string(cpp)
-        self.assertIn('constexpr int COUNT = 0;', writer.getvalue())
-
-    def test_is_constexpr_render_to_string_declaration(self):
-        writer = io.StringIO()
-        cpp = CppFile(None, writer=writer)
-        variables = CppVariable(name="COUNT",
-                                type="int",
-                                is_class_member=True,
-                                is_constexpr=True,
-                                initialization_value='0')
-        variables.render_to_string_declaration(cpp)
-        self.assertIn('constexpr int COUNT = 0;', writer.getvalue())
-
-    def test_is_extern_raises_error_when_is_static_is_true(self):
-        self.assertRaises(RuntimeError, CppVariable, name="var1", type="char*", is_static=True, is_extern=True)
-
-    def test_is_extern_render_to_string(self):
-        writer = io.StringIO()
-        cpp = CppFile(None, writer=writer)
-        v = CppVariable(name="var1", type="char*", is_extern=True)
-        v.render_to_string(cpp)
-        self.assertIn('extern char* var1;', writer.getvalue())
-
-
-class TestCppGenerator(unittest.TestCase):
-    """
-    Test C++ code generation
-    """
-
-    def test_cpp_variables(self):
-        generate_var(output_dir='.')
-        expected_cpp = ['var.cpp']
-        self.assertEqual(filecmp.cmpfiles('.', 'tests', expected_cpp)[0], expected_cpp)
-        os.remove(expected_cpp[0])
-
-    def test_cpp_arrays(self):
-        generate_array(output_dir='.')
-        expected_cpp = ['array.cpp']
-        self.assertEqual(filecmp.cmpfiles('.', 'tests', expected_cpp)[0], expected_cpp)
-        os.remove(expected_cpp[0])
-
-    def test_cpp_function(self):
-        generate_func(output_dir='.')
-        expected_cpp = ['func.cpp']
-        expected_h = ['func.h']
-        self.assertEqual(filecmp.cmpfiles('.', 'tests', expected_cpp)[0], expected_cpp)
-        self.assertEqual(filecmp.cmpfiles('.', 'tests', expected_h)[0], expected_h)
-        os.remove(expected_cpp[0])
-        os.remove(expected_h[0])
-
-    def test_cpp_enum(self):
-        generate_enum(output_dir='.')
-        expected_cpp = ['enum.cpp']
-        self.assertEqual(filecmp.cmpfiles('.', 'tests', expected_cpp)[0], expected_cpp)
-        os.remove(expected_cpp[0])
-
-    def test_cpp_class(self):
-        generate_class(output_dir='.')
-        expected_cpp = ['class.cpp']
-        expected_h = ['class.h']
-        self.assertEqual(filecmp.cmpfiles('.', 'tests', expected_cpp)[0], expected_cpp)
-        self.assertEqual(filecmp.cmpfiles('.', 'tests', expected_h)[0], expected_h)
-        os.remove(expected_cpp[0])
-        os.remove(expected_h[0])
 
 
 # Generate test data
